better Task & List cells
